add mouse interface for zooming in
This breaks when zooming in too much.  My theory here is the loss of
precision when scale is very small. [-2,2] seems to work OK

But digging down into smaller values results in all red screen

There might also be a bug there. Marking a selection starting in the
middle of the big blob, w/ x0,y0 inside bounds of it, the top-left of
the resulting body is not completely red as expected

_probably_ have a bug here with scaling.  Might still be floating point
shenanigans... hard to say, will dig into it later